Put all vvplay_async related struct inside src/vvplay_async_prefetch in future.
Moved NetworkTrace
